added MultiFetcherValueWrapper and FetcherResponseFactory
