+++ src/main/java/com/rentworthy/fetcher/MultiFetcher.java
@@ -1,0 +1,11 @@
+/**
+ * @author  cadams2
+ * @since   Feb 9, 2017
+ */
+package com.rentworthy.fetcher;
+
+import com.rentworthy.fetcher.response.FetcherResponse;
+
+public interface MultiFetcher<T> extends Fetcher<FetcherResponse<T>> {
+
+}
