--- src/main/java/com/rentworthy/fetcher/response/MultiFetcher.java
@@ -1,11 +1,0 @@
-/**
- * @author  cadams2
- * @since   Feb 9, 2017
- */
-package com.rentworthy.fetcher.response;
-
-import com.rentworthy.fetcher.Fetcher;
-
-public interface MultiFetcher<T> extends Fetcher<FetcherResponse<T>> {
-
-}
